cleaned up warnings, split test programs and libraries into client and server
