+++ openoffload/cpp/framework/src/opof_config.c
@@ -6,6 +6,7 @@
 #include <libconfig.h>
 
 #include <net/if.h>
+#include <arpa/inet.h>
 
 #include "opof.h"
 #include "opof_error.h"
@@ -14,20 +15,15 @@
 //void display_session_request(sessionRequest_t *request);
 
 sessionRequest_t **read_config(char *filename){
-	int status = 0;
   	sessionRequest_t *request;
 	sessionRequest_t **sessionRequests = NULL;
 	config_t cfg;
   	config_setting_t *requests;
   	const char *str;
-
+	int i;
+	int value;
+	const char *srcaddr, *dstaddr, *nexthopaddr;
   	config_init(&cfg);
-
-  	const char * interface_name;
-  	const char * type;
-  	int mtu;
-  	int frames, number,length;
-  	int version;
 
   	/* Read the file. If there is an error, report it and exit. */
   	if (! config_read_file(&cfg, filename))
@@ -35,7 +31,7 @@ sessionRequest_t **read_config(char *filename){
     	fprintf(stderr, "%s:%d - %s\n", config_error_file(&cfg),
             config_error_line(&cfg), config_error_text(&cfg));
     	config_destroy(&cfg);
-    	return(EXIT_FAILURE);
+    	return NULL;
   	}
 
 	/* Get the store name. */
@@ -50,9 +46,7 @@ sessionRequest_t **read_config(char *filename){
 
 	    int count = config_setting_length(requests);
 	    sessionRequests = (sessionRequest_t **)malloc(count * (sizeof(requests)));
-	    int i;
-	    int value;
-	    char *addr;
+
 	 	for(i = 0; i < count; ++i){
 	 		request = (sessionRequest_t *)malloc(sizeof(*request));
 	      	config_setting_t *session = config_setting_get_elem(requests, i);
@@ -65,20 +59,20 @@ sessionRequest_t **read_config(char *filename){
 		    request->inlif = value;
 		    config_setting_lookup_int(session, "outlif",&value);
 		    request->outlif = value;
-		    config_setting_lookup_string(session, "nexthop",&addr);
-		    inet_pton(AF_INET,addr, &value);
+		    config_setting_lookup_string(session, "nexthop",&nexthopaddr);
+		    inet_pton(AF_INET,nexthopaddr, &value);
 		    request->nextHop = value;
 		    config_setting_lookup_int(session, "protocolid",&value);
 		    request->proto = value;
 		    config_setting_lookup_int(session, "ipversion",&value);
 		    request->ipver = value;
-		    config_setting_lookup_string(session, "sourceip",&addr);
-		    inet_pton(AF_INET, addr, &value);
+		    config_setting_lookup_string(session, "sourceip",&srcaddr);
+		    inet_pton(AF_INET, srcaddr, &value);
 		    request->srcIP = value;
 		    config_setting_lookup_int(session, "sourceport",&value);
 		    request->srcPort = value;
-		    config_setting_lookup_string(session, "destinationip",&addr);
-		    inet_pton(AF_INET,addr, &value);
+		    config_setting_lookup_string(session, "destinationip",&dstaddr);
+		    inet_pton(AF_INET,dstaddr, &value);
 		    request->dstIP = value;
  			config_setting_lookup_int(session, "destinationport",&value);
 		    request->dstPort = value;
