+++ openoffload/cpp/framework/src/opof_util.cc
@@ -20,7 +20,7 @@ extern "C" {
 
 #include "opof_util.h"
 #include "opof_grpc.h"
-#include "opof_session_client.h"
+//#include "opof_session_client.h"
 
 
 
@@ -37,33 +37,7 @@ int create_address(char *address, unsigned short port){
 }
 
 
-void display_session_response(sessionResponse_t *response){
 
-    printf("\n\nSession Response\n");
-    printf("Session ID: %ld\n",response->sessionId);
-    printf("In Packets %ld\n",response->inPackets);
-    printf("Out Packets: %ld\n",response->outPackets);
-    printf("In Bytes: %ld\n",response->inBytes);
-    printf("Out Bytes: %ld\n",response->outBytes);
-    printf("Session State: %d\n",response->sessionState);
-    printf("Session Close Code; %d\n",response->sessionCloseCode);
-    printf("Request Status: %d\n",response->requestStatus);
-}
-
-void display_session_request(sessionRequest_t *request){
-
-    printf("\n\nSession Request\n");
-    printf("Session ID: %ld\n",request->sessId);
-    printf( "Inlif: %d\n",request->inlif);
-    printf( "Outlif: %d\n",request->outlif);
-    printf( "Source Port: %d\n",request->srcPort);
-    printf( "Source IP: %d\n", request->srcIP);
-    printf( "Destination IP: %d\n",request->dstIP);
-    printf( "Destination Port: %d\n",request->dstPort);
-    printf( "Protocol ID: %d\n",request->proto);
-    printf( "IP Version: %d\n",request->ipver);
-    printf( "Action Value: %d\n",request->actType);
-}
 
 void convertSessionRequest2cpp(sessionRequest_t *request_c, sessionRequest *request){
     actionParameters action;
